Add license headers, move some packets to more appropriate namespaces
